avoid x86_64 warning
svn: r11346

+++ src/mzscheme/src/jit.c
@@ -7162,7 +7162,7 @@ static void on_demand_generate_lambda(Scheme_Native_Closure *nc)
     void **pd;
     pd = (void **)gdata.patch_depth;
     gdata.patch_depth = pd[1];
-    jit_patch_movi((*pd), (void *)max_depth);
+    jit_patch_movi((*pd), (void *)(long)max_depth);
   }
 
   ndata->code = code;
